refactor: restructure with subcommands using Cobra
- Created cmd/ package structure for better organization
- Split functionality into separate subcommands:
  - cmd/root.go: main command and coordination
  - cmd/version.go: version subcommand
  - cmd/branches.go: branch cleanup logic (default command)
- Simplified main.go to just call cmd.Execute()
- Default behavior: runs 'branches' when no subcommand specified
- Commands now work as:
  - git-gone (runs branches by default)
  - git-gone version
  - git-gone branches
  - git-gone help
- Improved help output with proper command hierarchy
- Ready for future subcommand extensions
