+++ cmd/root.go
@@ -1,0 +1,41 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+// Version information - set during build with ldflags
+var (
+	Version    = "0.2.0"
+	CommitHash = "unknown"
+	BuildTime  = "unknown"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "git-gone",
+	Short: "Clean up merged git branches interactively",
+	Long: `git-gone - Clean up merged git branches interactively
+
+git-gone helps you clean up local git branches that have been merged
+or whose remote tracking branches have been deleted.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		// By default, run the branches command when no subcommand is provided
+		branchesCmd.Run(cmd, args)
+	},
+}
+
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Fprintln(os.Stderr, err)
+		os.Exit(1)
+	}
+}
+
+func init() {
+	// Add subcommands
+	rootCmd.AddCommand(branchesCmd)
+	rootCmd.AddCommand(versionCmd)
+}
