+++ cmd/version.go
@@ -1,0 +1,18 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Short: "Show version information",
+	Long:  `Display the version, commit hash, and build time of git-gone.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Printf("git-gone %s\n", Version)
+		fmt.Printf("Commit: %s\n", CommitHash)
+		fmt.Printf("Built: %s\n", BuildTime)
+	},
+}
