+++ go.mod
@@ -2,7 +2,10 @@ module git-gone
 
 go 1.19
 
-require github.com/koki-develop/go-fzf v0.15.0
+require (
+	github.com/koki-develop/go-fzf v0.15.0
+	github.com/spf13/cobra v1.10.1
+)
 
 require (
 	github.com/atotto/clipboard v0.1.4 // indirect
@@ -21,7 +24,6 @@ require (
 	github.com/muesli/reflow v0.3.0 // indirect
 	github.com/muesli/termenv v0.15.2 // indirect
 	github.com/rivo/uniseg v0.2.0 // indirect
-	github.com/spf13/cobra v1.10.1 // indirect
 	github.com/spf13/pflag v1.0.10 // indirect
 	golang.org/x/sync v0.3.0 // indirect
 	golang.org/x/sys v0.7.0 // indirect
